Copy RoleDaoEjb to RoleDaoJpa. Changed RoleDaoEjb and RoleDaoJpa.
